refactor Sidebar
